src/seminars/first/Shop/Shop.java
package seminars.first.Shop;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

public class Shop {
    private List<Product> products;

    // Геттеры, сеттеры:
    public List<Product> getProducts() {
        return products;
    }

    public void setProducts(List<Product> products) {
        this.products = products;
    }

    // Метод должен вернуть отсортированный по возрастанию по цене список продуктов
    public List<Product> sortProductsByPrice() {
        // Допишите реализацию метода самостоятельно
        List<Product> sortedList = new ArrayList<>(products);
        Collections.sort(sortedList);
        return sortedList;
    }

    // Метод должен вернуть самый дорогой продукт
    public Product getMostExpensiveProduct() {
        // Допишите реализацию метода самостоятельно
        List<Product> sortedList = new ArrayList<>(products);
        Collections.sort(sortedList);
        return sortedList.get(sortedList.size() - 1);
    }
}